+++ processor/src/trace/utils.rs
@@ -130,7 +130,7 @@ pub trait AuxColumnBuilder<H: Copy, R: LookupTableRow> {
 
     /// Returns a sequence of hints which indicate how the table was updated. Each hint consists
     /// of a clock cycle at which the update happened as well as the hint describing the update.
-    fn get_table_hints(&self) -> &[(usize, H)];
+    fn get_table_hints(&self) -> &[(u32, H)];
 
     /// Returns a value by which the current value of the column should be multiplied to get the
     /// next value. It is expected that this value should never be ZERO in practice.
@@ -157,11 +157,11 @@ pub trait AuxColumnBuilder<H: Copy, R: LookupTableRow> {
         result[0] = self.init_column_value(alphas);
 
         // keep track of the last updated row in the running product column
-        let mut result_idx = 0;
+        let mut result_idx = 0_usize;
 
         // iterate through the list of updates and apply them one by one
         for (clk, hint) in self.get_table_hints() {
-            let clk = *clk;
+            let clk = *clk as usize;
 
             // if we skipped some cycles since the last update was processed, values in the last
             // updated row should by copied over until the current cycle.
